requirements.txt
Django==2.2.2
django-debug-toolbar==2.0
django-richtextfield==1.4.0
django-session-timeout==0.0.4
Pillow==6.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
